requirements.txt
Django==1.7
djangorestframework==3.1.1
pytest==2.7.0
wrapt==1.10.2
django_extensions==1.5.2
